Create Q1. Explain what the simple List component does.
In React Js, Lists are used to display data in an ordered format and mainly used to display menus on websites. In React, Lists can be created in a similar way as we create lists in JavaScript.The List components in React gives you a step by step walkthrough on how to render a list of simple primitives, how to render a list of complex objects, and how to update the state of your list in React.

There are various List components available in react, Above is an example code for a basic list in react js.